+++ PracticaLink/backend/app/main.py
@@ -4,6 +4,8 @@ from sqlalchemy import text
 
 from app.core.database import engine
 from app.api.routes.auth import router as auth_router
+from app.api.routes.practicas import router as practicas_router
+from app.api.routes.estudiantes import router as estudiantes_router
 
 app = FastAPI(
     title="PracticaLink API",
@@ -23,6 +25,8 @@ app.add_middleware(
 )
 
 app.include_router(auth_router)
+app.include_router(practicas_router)
+app.include_router(estudiantes_router)
 
 
 @app.get("/")